Django/data_visualization/gdp_analysis/requirements.txt
bokeh
django-extensions
django-htmx